+++ 4-full-mvc-framework/config/config.php
@@ -32,6 +32,17 @@ define('AVATAR_PATH', 'public/avatars/'); // dont forget to make this folder wri
  */
 define('USE_GRAVATARS', false);
 
+/**
+ * Configuration for: Cookies
+ * Please note: The COOKIE_DOMAIN needs the domain where your app is, 
+ * in a format like this: .mydomain.com
+ * Note the . in front of the domain. No www, no http, no slash here!
+ * For local development .127.0.0.1 is fine, but when deploying you should
+ * change this to your real domain!
+ * @see php.net/manual/en/function.setcookie.php
+ */
+define('COOKIE_RUNTIME', 1209600); // 1209600 seconds = 2 weeks
+define('COOKIE_DOMAIN', '.127.0.0.1'); // the domain where the cookie is valid for
 
 /**
  * Configuration for: Database